Changed structure of classes
